API closer to json/pickle
The json and pickle libraries use dump and dumps for writing to a file and str, respectively.
The bconv formatters had write and dump, which are now renamed to write and dumps
to avoid confusion.
Also, the arguments to write() are flipped: (stream, content) -> (content, stream) to be
conformant with `json.dump(obj, stream)`.
The write() method is not renamed to dump() in order to keep the current level of clarity
in the code (dump and dumps are too close).
There is a dump() method at the abstract base class, however, which is an alias for write().